Solution to Count Students Scoring Above a Threshold

+++ Count Students Scoring Above a Threshold@@67877d61dc107486ad5c2747/code.c
@@ -10,7 +10,7 @@ int main (){
         scanf("%d %s %f",&rollno[i],name[i],&marks[i]);
     }
 
-    int limit ;
+    float limit ;
     int count = 0 ;
     scanf("%f",&limit);
